+++ src/execute.rs
@@ -256,35 +256,6 @@ impl ExecuteCtx {
         let config = &configure_wasmtime(wasm_features, profiling.native_strategy());
         let engine = Engine::new(config)?;
         let instance_pre = if is_component {
-            warn!(
-                "
-
-   +------------------------------------------------------------------------+
-   |                                                                        |
-   | Wasm Component support in viceroy is in active development, and is not |
-   |                    supported for general consumption.                  |
-   |                                                                        |
-   +------------------------------------------------------------------------+
-
-            "
-            );
-
-            // If logging isn't enabled, print the notice to stderr.
-            if !tracing::enabled!(Level::WARN) {
-                eprintln!(
-                    "
-
-   +------------------------------------------------------------------------+
-   |                                                                        |
-   | Wasm Component support in viceroy is in active development, and is not |
-   |                    supported for general consumption.                  |
-   |                                                                        |
-   +------------------------------------------------------------------------+
-
-            "
-                );
-            }
-
             let mut linker: component::Linker<ComponentCtx> = component::Linker::new(&engine);
             compute::link_host_functions(&mut linker)?;
             let component = if is_wat {